+++ features/MiniFarms/MiniFarmCollapsible.tsx
@@ -14,7 +14,7 @@ import { ERC20Transfer } from "../../containers/Erc20Transfer";
 import Collapse from "../Collapsible/Collapse";
 import { JarApy } from "../../containers/Jars/useJarsWithAPYPFCore";
 import { LpIcon, TokenIcon, MiniIcon } from "../../components/TokenIcon";
-import { useButtonStatus, ButtonStatus } from "hooks/useButtonStatus";
+import { useButtonStatus, ButtonStatus } from "v1/hooks/useButtonStatus";
 import { PickleCore } from "containers/Jars/usePickleCore";
 
 interface DataProps {
