+++ v1/hooks/useGaugeProxy.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from "ethers";
 import { useState } from "react";
-import { Contracts } from "../containers/Contracts";
+import { Contracts } from "../../containers/Contracts";
 
 export enum TransactionStatus {
   Pending,
